+++ src/App.js
@@ -1,39 +1,13 @@
 import * as React from 'react';
+import TodoList from './TodoList';
+import AddTodoForm from './AddTodoForm';
 
-const todoList = [
-  {
-    id: 1,
-    title: 'Complete assignment',
-  },
-  {
-    id: 2,
-    title: 'Eat',
-  },
-  {
-    id: 3,
-    title: 'Sleep',
-  },
-  {
-    id: 3,
-    title: 'Repeat',
-  },
-
-];
-
-function App() {
-  return (
-    <div>
-      <h1>Todo List</h1>
-
-      <hr />
- 
-      <ul>
-        {todoList.map(function (item) {
-          return <li key={item.id}>{item.title}</li>;
-        })}
-      </ul>
-    </div>
-  );
-}
+let App= () => (
+  <div>
+    <h1>Todo List</h1>
+    <AddTodoForm />
+    <TodoList />
+  </div>
+)
  
 export default App;
